Added caching and error handler to useQuizQuery hook
